refactor(session): 타로 세션·설정 페이지 컨트롤러 훅 분리 (B-5 저위험)
대형 페이지를 view/controller로 분리 — 로직을 훅으로 verbatim 이동해 동작·DOM·
E2E 셀렉터를 보존(저위험). 레이스 컨디션 가드(ref-lock·타이머 중복차단·TOCTOU)는 원본 그대로.

- 타로 세션: 카드선택·리딩진행 로직 → useTarotCardSelection 훅. page 532→257줄
- 설정: 상태·핸들러·storage 로직 → useSettings 훅. page 350→241줄
- 신규 훅 2개 sonar.coverage/cpd.exclusions 동기화, CLAUDE.md 훅 목록 갱신

검증: type-check·lint(0 errors)·build 성공·984 테스트·커버리지 게이트 통과.
DOM/JSX 불변이라 E2E 셀렉터 영향 0 (CI E2E로 동작 최종 검증). 마이그레이션 0.
잔여(후속): saju(331)·shinjeom(411) 세션 페이지 동일 패턴 분리.
