+++ sonar-project.properties
@@ -56,6 +56,8 @@ sonar.coverage.exclusions=\
   src/lib/guest-sessions.ts,\
   src/lib/particle-engine.ts,\
   src/hooks/useTarotReading.ts,\
+  src/hooks/useTarotCardSelection.ts,\
+  src/hooks/useSettings.ts,\
   src/hooks/useMouseParallax.ts,\
   src/components/effects/ServiceIllustrations.tsx,\
   src/lib/supabase/client.ts,\
@@ -113,6 +115,8 @@ sonar.cpd.exclusions=\
   src/components/tarot/CardInterpretationList.tsx,\
   src/components/tarot/TarotResultPanel.tsx,\
   src/hooks/useTarotReading.ts,\
+  src/hooks/useTarotCardSelection.ts,\
+  src/hooks/useSettings.ts,\
   src/components/saju/SajuChartReveal.tsx,\
   src/components/shinjeom/ShinjeomEnergyEffect.tsx,\
   src/components/session/ResultTextCard.tsx,\
